Refactor to use fixed-point arithmetic (Fxp) for various parameters and calculations
- Updated BulletMLState to use Fxp instead of double for parameters.
- Modified bulletHitsShip function to use integer types and long long for calculations, improving precision and performance.
- Changed Foe structure to use Fxp for rank and updated related functions to accept Fxp.
- Refactored FoeCommand methods to utilize Fxp for direction, speed, and other calculations.
- Adjusted letter rendering to use Fxp for coordinates and sizes, enhancing rendering precision.
- Replaced float arrays in letter data with a structured LetterStroke using Fxp.
- Updated stage parameters to use a struct with Fxp for startLevel and levelInc.
- Enhanced vector operations to return long long for inner product calculations, ensuring better precision.

+++ src/vector.cpp
@@ -39,10 +39,10 @@ static int intSqrtU64(unsigned long long value)
   return (int)res;
 }
 
-float vctInnerProduct(Vector *v1, Vector *v2) {
+long long vctInnerProduct(Vector *v1, Vector *v2) {
   const long long x = (long long)v1->x * (long long)v2->x;
   const long long y = (long long)v1->y * (long long)v2->y;
-  return (float)(x + y);
+  return x + y;
 }
 
 Vector vctGetElement(Vector *v1, Vector *v2) {
